Enhance landing page with scroll animations and interactive design elements
